[refactor] Landing: 일러스트 컴포넌트 export용 index.ts 파일 생성
