+++ src/app/(landing)/_components/TopSection/Illustration/index.ts
